Função que calcula convergência no ambiente movida para a camada de visualização
